I couldn't find an email address
Apologies for unsolicited PR(I would have opened an issue but it seemed un polite)  Feel free to close this, I couldn't find the email on golangprojects.com ( a job board with the job posting about a backend position). So, I updated this repo README which seemed to have recent activity with a summary of who I am and why I am interested in the position.